% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Debug_Assert}
\alias{Debug_Assert}
\alias{module_a_func0}
\alias{module_a_func1}
\alias{module_a_func1b}
\alias{module_a_func2}
\alias{module_a_func3}
\alias{module_a_func4}
\title{Debug_Assert}
\usage{
module_a_func0()

module_a_func1()

module_a_func1b()

module_a_func2()

module_a_func3()

module_a_func4()
}
\value{
All functions have a default return of \code{TRUE}
}
\description{
Illustration of \sQuote{Debug_Assert}
}
